% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{get_recipe}
\alias{get_recipe}
\title{This function gives the summary to a specific recipe-id gathered from get_recommendations.}
\usage{
get_recipe(id, key)
}
\arguments{
\item{id}{recipe id}

\item{key}{app credentials from rapidapi}
}
\value{
Returns a single character string summarizing the specific recipe
}
\description{
This function gives the summary to a specific recipe-id gathered from get_recommendations.
}
