% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendations.R
\name{get_recommendations}
\alias{get_recommendations}
\title{This function gives the summary to a specific recipe-id gathered from get_recommendations.}
\usage{
get_recommendations(ingredients, no_results, key)
}
\arguments{
\item{ingredients}{a list of characters, specifying the ingredients}

\item{no_results}{integer specifying the number of recommendations}

\item{key}{app credentials from rapidapi}
}
\value{
Returns a dataframe, which recommends recipes based on the specified ingredients.
}
\description{
This function gives the summary to a specific recipe-id gathered from get_recommendations.
}
